Send a unique identifier instead of email for OIDC sub
Currently the system sends the user's current email address
to SSO applications as the subject identifier. This means that
if the user's email were to change, they would appear as
a new user to SSO clients.

Resolves this issue by instead creating and sending an unchanging
UUID as the sub. For legacy users their original registration
email is used instead in order to not break compatibility with
existing SSO signins.