Allow users to change their own email and password.
Adds fields on the "Change my settings" page that allow users to change
their email and password. (Password changes were already possible but
required using the "Forgot password" flow).

For password changes, an email notification is sent informing the user
that their password was changed.

For email changes, a confirmation email is sent to the new email to
guard against accidental changes to invalid email addresses. After
the change a notification is also sent to the old email.